requirements.txt
asgiref==3.3.4
Django==3.2
django-autoslug==1.9.8
django-extensions==3.1.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
